+++ core/main.go
@@ -25,6 +25,7 @@ import (
 	"github.com/mesg-foundation/engine/config"
 	"github.com/mesg-foundation/engine/container"
 	"github.com/mesg-foundation/engine/cosmos"
+	"github.com/mesg-foundation/engine/errors"
 	"github.com/mesg-foundation/engine/hash"
 	"github.com/mesg-foundation/engine/logger"
 	"github.com/mesg-foundation/engine/orchestrator"
@@ -33,10 +34,6 @@ import (
 	runnersdk "github.com/mesg-foundation/engine/sdk/runner"
 	"github.com/mesg-foundation/engine/server/grpc"
 	"github.com/mesg-foundation/engine/version"
-	"github.com/mesg-foundation/engine/x/xerrors"
-	"github.com/mesg-foundation/engine/x/xnet"
-	"github.com/mesg-foundation/engine/x/xrand"
-	"github.com/mesg-foundation/engine/x/xsignal"
 	"github.com/sirupsen/logrus"
 	rpcserver "github.com/tendermint/tendermint/rpc/lib/server"
 	tmtypes "github.com/tendermint/tendermint/types"
@@ -68,7 +65,7 @@ func stopRunningServices(sdk *enginesdk.SDK, address string) error {
 	}
 	wg.Wait()
 	close(errC)
-	var errs xerrors.Errors
+	var errs errors.Errors
 	for err := range errC {
 		errs = append(errs, err)
 	}
